requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
chromadb==0.4.22
sentence-transformers==2.3.1
pyautogui==0.9.54
psutil==5.9.8
python-dotenv==1.0.1
sqlalchemy==2.0.25
pydantic==2.6.0
pydantic-settings==2.1.0
numpy==1.26.3
scipy==1.12.0
pandas==2.2.0
requests==2.31.0
aiohttp==3.9.3
python-multipart==0.0.6
pillow==10.2.0
opencv-python==4.9.0.80

# OpenClaw - Web Scraping & Automation
beautifulsoup4==4.12.3
selenium==4.16.0
lxml==5.1.0
html5lib==1.1
sounddevice==0.4.6
soundfile==0.12.1
pyttsx3==2.90
pyaudio==0.2.14
openai-whisper==20231117
webrtcvad-wheels==2.0.10
pvporcupine==3.0.2
tiktoken==0.5.2
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
flake8==7.0.0
mypy==1.8.0
black==24.1.1

# Windows-specific for GUI automation
pywin32==306; sys_platform == 'win32'

# AI Provider SDKs
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
groq==0.4.1
litellm==1.30.0

# Security & Encryption
cryptography==42.0.2
joblib==1.3.2

# Machine Learning for Threat Detection
scikit-learn==1.4.0

# Multi-Agent Framework (optional - for advanced features)
# langchain==0.1.0
# langchain-openai==0.0.2

# Bug Bounty Automation
dnspython==2.5.0

# Document Intelligence
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23

# Multi-Language Support
langdetect==1.0.9

# Screen Understanding & OCR
pytesseract==0.3.10
easyocr==1.7.0

# Discord Bot Integration (Mekio-style)

# Natural Voice (Free)
edge-tts==6.1.10
nest-asyncio==1.6.0
